fix: date dependency corrections
